kedrSite/requirements.txt
asgiref==3.8.1
Django==4.2.20
django-appconf==1.1.0
djangorestframework==3.16.0
pillow==11.1.0
psycopg2==2.9.10
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
